+++ src/main/java/com/neuronrobotics/sdk/addons/kinematics/PidPrismaticLink.java
@@ -30,7 +30,7 @@ public class PidPrismaticLink extends AbstractPrismaticLink{
 	 */
 	@Override
 	public void cacheTargetValueDevice() {
-		channel.setCachedTargetValue(getTargetValue());
+		channel.setCachedTargetValue((int)getTargetValue());
 	}
 
 	/* (non-Javadoc)
@@ -54,7 +54,7 @@ public class PidPrismaticLink extends AbstractPrismaticLink{
 	 * @see com.neuronrobotics.sdk.addons.kinematics.AbstractLink#getCurrentPosition()
 	 */
 	@Override
-	public int getCurrentPosition() {
+	public double getCurrentPosition() {
 		int val=channel.GetPIDPosition();
 		fireLinkListener(val);
 		return val;
